requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
cloudinary==1.28.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.0
django-cors-headers==3.10.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.0.0
psycopg2==2.9.3
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
